+++ internal/db/vectorindex/errors.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package vectorstore
+package vectorindex
 
 import (
 	"github.com/sourcenetwork/defradb/errors"
